+++ memor/prompt.py
@@ -7,6 +7,7 @@ from .params import MEMOR_VERSION
 from .params import DATE_TIME_FORMAT
 from .params import RenderFormat, DATA_SAVE_SUCCESS_MESSAGE
 from .params import Role
+from .tokens_estimator import TokensEstimator
 from .params import INVALID_PROMPT_STRUCTURE_MESSAGE, INVALID_TEMPLATE_MESSAGE
 from .params import INVALID_ROLE_MESSAGE, INVALID_RESPONSE_MESSAGE
 from .params import PROMPT_RENDER_ERROR_MESSAGE
@@ -407,3 +408,11 @@ class Prompt:
             return True
         except Exception:
             return False
+
+    def estimate_tokens(self, method: TokensEstimator = TokensEstimator.DEFAULT) -> int:
+        """
+        Estimate the number of tokens in the prompt message.
+
+        :param method: token estimator method
+        """
+        return method(self.render(render_format=RenderFormat.STRING))