+++ memor/session.py
@@ -11,6 +11,7 @@ from .params import INVALID_MESSAGE_STATUS_LEN_MESSAGE
 from .params import INVALID_RENDER_FORMAT_MESSAGE
 from .params import UNSUPPORTED_OPERAND_ERROR_MESSAGE
 from .params import RenderFormat
+from .tokens_estimator import TokensEstimator
 from .prompt import Prompt
 from .response import Response
 from .errors import MemorValidationError
@@ -352,6 +353,14 @@ class Session:
             return True
         except Exception:
             return False
+
+    def estimate_tokens(self, method: TokensEstimator = TokensEstimator.DEFAULT) -> int:
+        """
+        Estimate the number of tokens in the session.
+
+        :param method: token estimator method
+        """
+        return method(self.render(render_format=RenderFormat.STRING))
 
     @property
     def date_created(self) -> datetime.datetime: