+++ tests/test_prompt.py
@@ -4,6 +4,7 @@ import pytest
 from memor import Prompt, Response, Role
 from memor import PresetPromptTemplate, PromptTemplate
 from memor import RenderFormat, MemorValidationError, MemorRenderError
+from memor import TokensEstimator
 
 TEST_CASE_NAME = "Prompt tests"
 
@@ -45,6 +46,16 @@ def test_tokens4():
     prompt = Prompt(message="Hello, how are you?", role=Role.USER)
     with pytest.raises(MemorValidationError, match=r"Invalid value. `tokens` must be a positive integer."):
         prompt.update_tokens("4")
+
+
+def test_estimated_tokens1():
+    prompt = Prompt(message="Hello, how are you?", role=Role.USER)
+    assert prompt.estimate_tokens(TokensEstimator.UNIVERSAL) == 7
+
+
+def test_estimated_tokens2():
+    prompt = Prompt(message="Hello, how are you?", role=Role.USER)
+    assert prompt.estimate_tokens(TokensEstimator.OPENAI) == 7
 
 
 def test_role1():